refactor: Moved modules to modules folder
