+++ src/app.module.ts
@@ -1,11 +1,11 @@
-import { ArtistModule } from './modules/artist/artist.module';
 import { Module } from '@nestjs/common';
 import { CoreModule } from './common/core.module';
 import { InMemoryDBService } from './in-memory-db';
+import { ArtistModule } from './modules/artist/artist.module';
 import { UsersModule } from './modules/users/users.module';
-import { AlbumModule } from './album/album.module';
-import { TrackModule } from './track/track.module';
-import { FavsModule } from './favs/favs.module';
+import { AlbumModule } from './modules/album/album.module';
+import { TrackModule } from './modules/track/track.module';
+import { FavsModule } from './modules/favs/favs.module';
 
 @Module({
   imports: [
