+++ src/modules/album/album.module.ts
@@ -1,4 +1,3 @@
-import { ArtistService } from './../modules/artist/artist.service';
 import { Module } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
@@ -6,6 +5,6 @@ import { InMemoryDBService } from 'src/in-memory-db';
 
 @Module({
   controllers: [AlbumController],
-  providers: [AlbumService, ArtistService, InMemoryDBService]
+  providers: [AlbumService, InMemoryDBService]
 })
 export class AlbumModule { }
