+++ src/modules/favs/entities/fav.entity.ts
@@ -1,6 +1,6 @@
-import { AlbumEntity } from "src/album/entities/album.entity";
+import { AlbumEntity } from "src/modules/album/entities/album.entity";
 import { ArtistEntity } from "src/modules/artist/entities/artist.entity";
-import { TrackEntity } from "src/track/entities/track.entity";
+import { TrackEntity } from "src/modules/track/entities/track.entity";
 
 export interface FavsEntity {
   artists: ArtistEntity[];
